Descriptions for rest of validators
Wrote up descriptions with examples for remaining validators.

I think some could do with a rename but am not doing this for now as would require co-ordinated refactor, which will be easier once Pay JS commons is actually in use.

But for example `passwordLessThanTenChars` to me sounds like its checking if a password is less than 10 but its checking the opposite of that